Add useSingleSelectListState
